AAGSB-1174 add  class definition
