+++ package.xml
@@ -21,13 +21,15 @@
   <build_depend>async_web_server_cpp</build_depend>
   <build_depend>ffmpeg</build_depend>
   <build_depend>sensor_msgs</build_depend>
-
+  <build_depend>rclcpp_components</build_depend>
+  
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>cv_bridge</exec_depend>
   <exec_depend>image_transport</exec_depend>
   <exec_depend>async_web_server_cpp</exec_depend>
   <exec_depend>ffmpeg</exec_depend>
   <exec_depend>sensor_msgs</exec_depend>
+  <exec_depend>rclcpp_components</exec_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
